refactor query builders but pulling out common logic into a base class
