+++ app/services/search/query_builders/classified_listing.rb
@@ -1,6 +1,6 @@
 module Search
   module QueryBuilders
-    class ClassifiedListing
+    class ClassifiedListing < QueryBase
       TERM_KEYS = %i[
         category
         contact_via_connect
@@ -28,8 +28,6 @@ module Search
         size: 0
       }.freeze
 
-      attr_accessor :params, :body
-
       def initialize(params)
         @params = params.deep_symbolize_keys
 
@@ -41,36 +39,12 @@ module Search
         build_body
       end
 
-      def as_hash
-        @body
-      end
-
       private
-
-      def build_body
-        @body = ActiveSupport::HashWithIndifferentAccess.new
-        build_queries
-        add_sort
-        set_size
-      end
 
       def build_queries
         @body[:query] = { bool: {} }
         @body[:query][:bool][:filter] = filter_conditions
         @body[:query][:bool][:must] = query_conditions if query_keys_present?
-      end
-
-      def add_sort
-        sort_key = @params[:sort_by] || DEFAULT_PARAMS[:sort_by]
-        sort_direction = @params[:sort_direction] || DEFAULT_PARAMS[:sort_direction]
-        @body[:sort] = {
-          sort_key => sort_direction
-        }
-      end
-
-      def set_size
-        # By default we will return 0 documents if size is not specified
-        @body[:size] = @params[:size] || DEFAULT_PARAMS[:size]
       end
 
       def filter_conditions
@@ -103,25 +77,6 @@ module Search
 
       def range_keys_present?
         RANGE_KEYS.detect { |key| @params[key].present? }
-      end
-
-      def query_keys_present?
-        QUERY_KEYS.detect { |key| @params[key].present? }
-      end
-
-      def query_conditions
-        QUERY_KEYS.map do |query_key|
-          next if @params[query_key].blank?
-
-          {
-            simple_query_string: {
-              query: "#{@params[query_key]}*",
-              fields: [query_key],
-              lenient: true,
-              analyze_wildcard: true
-            }
-          }
-        end.compact
       end
     end
   end
